feature: add ExM shared object and import script
- Add the ExM shared object
This acts as a mirror to the old ESX object through the use of metatables however, allows the ExM object to override the ESX object without effecting any backwards compatibility and doesn't allow access to new ExM functions from the ESX object.

- Add imports.lua
This is the way to access the new ExM shared object in resources. Simply import this file within your resource's fxmanifest (shared_script "@extendedmode/imports.lua") and the ExM object will be present.