Dev (#1)
* compat (server) - Add getBank compat function

- Some older version of esx used a getBank function, adding it back for compat purposes.

* compat (server) - Add removeBank and addBank compat function

- Also add removeBank for older esx versions for compatibility
- And same for addBank

* Updated br.lua

Fixed some missing accents, translated what needed translations. At Weapon Ammo and Weapon Skin some stuff might not be translated as i dont know the exact translation to Brazilian Portuguese (Im European)

* Removed a lost - in br.lua

title

* Update br.lua

* tweak (server) - Switch to a cfg option for first spawn

This is a revised version of #82 by JnKTechstuff

* Fix issues when removing some vehicle properties (#111)

* Fix issues when removing some vehicle properties

There are some issues with esx_lscustom, when you hover on turbo or xenon (and I think smoke), they are applied but not removed, this fixed it for me. This should also work if you want to remove those properties from the vehicle.

* Rewrite: Updates to es.lua #96 [CalipoPA]

Co-authored-by: Smallo <smallo92@gmail.com>
Co-authored-by: HypeLevels <61633775+HypeLevels@users.noreply.github.com>
Co-authored-by: Neddings <github@neddings.com>
Co-authored-by: Alexander Schilling <19366689+alexander-schilling@users.noreply.github.com>
Co-authored-by: Neddings <53919357+Neddings@users.noreply.github.com>